+++ java/src/mihai/camera/ImageSaver.java
@@ -123,7 +123,7 @@ public class ImageSaver extends Thread implements ImageReader.Listener
         }
     }
 
-    public void handleImage(byte[] im, ImageSourceFormat ifmt, double time, int camera)
+    public void handleImage(byte[] im, ImageSourceFormat ifmt, long time, int camera)
     {
         synchronized(lock)
         {
